+++ src/flask_sqlalchemy/extension.py
@@ -4,9 +4,7 @@ import warnings
 from threading import Lock
 
 import sqlalchemy
-from flask import abort
 from flask import current_app
-from flask import request
 from sqlalchemy import event
 from sqlalchemy import orm
 from sqlalchemy.engine.url import make_url
@@ -15,7 +13,7 @@ from sqlalchemy.orm.session import Session as SessionBase
 
 from .model import DefaultMeta
 from .model import Model
-from .pagination import Pagination
+from .query import BaseQuery
 from .record_queries import _EngineDebuggingSignalEvents
 from .track_modifications import _SessionSignalEvents
 
@@ -153,111 +151,6 @@ class SignallingSession(SessionBase):
                 return state.db.get_engine(self.app, bind=bind_key)
 
         return super().get_bind(mapper, **kwargs)
-
-
-class BaseQuery(orm.Query):
-    """SQLAlchemy :class:`~sqlalchemy.orm.query.Query` subclass with
-    convenience methods for querying in a web application.
-
-    This is the default :attr:`~Model.query` object used for models, and
-    exposed as :attr:`~SQLAlchemy.Query`. Override the query class for
-    an individual model by subclassing this and setting
-    :attr:`~Model.query_class`.
-    """
-
-    def get_or_404(self, ident, description=None):
-        """Like :meth:`get` but aborts with 404 if not found instead of
-        returning ``None``.
-        """
-        rv = self.get(ident)
-        if rv is None:
-            abort(404, description=description)
-        return rv
-
-    def first_or_404(self, description=None):
-        """Like :meth:`first` but aborts with 404 if not found instead
-        of returning ``None``.
-        """
-        rv = self.first()
-        if rv is None:
-            abort(404, description=description)
-        return rv
-
-    def paginate(
-        self, page=None, per_page=None, error_out=True, max_per_page=None, count=True
-    ):
-        """Returns ``per_page`` items from page ``page``.
-
-        If ``page`` or ``per_page`` are ``None``, they will be retrieved from
-        the request query. If ``max_per_page`` is specified, ``per_page`` will
-        be limited to that value. If there is no request or they aren't in the
-        query, they default to 1 and 20 respectively. If ``count`` is ``False``,
-        no query to help determine total page count will be run.
-
-        When ``error_out`` is ``True`` (default), the following rules will
-        cause a 404 response:
-
-        * No items are found and ``page`` is not 1.
-        * ``page`` is less than 1, or ``per_page`` is negative.
-        * ``page`` or ``per_page`` are not ints.
-
-        When ``error_out`` is ``False``, ``page`` and ``per_page`` default to
-        1 and 20 respectively.
-
-        Returns a :class:`Pagination` object.
-        """
-
-        if request:
-            if page is None:
-                try:
-                    page = int(request.args.get("page", 1))
-                except (TypeError, ValueError):
-                    if error_out:
-                        abort(404)
-
-                    page = 1
-
-            if per_page is None:
-                try:
-                    per_page = int(request.args.get("per_page", 20))
-                except (TypeError, ValueError):
-                    if error_out:
-                        abort(404)
-
-                    per_page = 20
-        else:
-            if page is None:
-                page = 1
-
-            if per_page is None:
-                per_page = 20
-
-        if max_per_page is not None:
-            per_page = min(per_page, max_per_page)
-
-        if page < 1:
-            if error_out:
-                abort(404)
-            else:
-                page = 1
-
-        if per_page < 0:
-            if error_out:
-                abort(404)
-            else:
-                per_page = 20
-
-        items = self.limit(per_page).offset((page - 1) * per_page).all()
-
-        if not items and page != 1 and error_out:
-            abort(404)
-
-        if not count:
-            total = None
-        else:
-            total = self.order_by(None).count()
-
-        return Pagination(self, page, per_page, total, items)
 
 
 class _QueryProperty:
