+++ tests/test_query_class.py
@@ -1,5 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
-from flask_sqlalchemy.extension import BaseQuery
+from flask_sqlalchemy.query import BaseQuery
 
 
 def test_default_query_class(db):
